Extracting gif cache
- moved it in different file
